+++ NEWS.md
@@ -1,5 +1,6 @@
 # scsf dev
 
+* replace sc_rand with sc_uid from sc
 
 * several updates to better support the sc workers and sf round-trip